Added sync.sql to keep backups in sync (without using rsync)
When files/folders are renmaed, rsync deletes and re-copies the files. This can be
particularly problematic for large (video) files. Hence, this script will look
at the file_audit_log table and generate the necessary 'bash' commands to keep
the backups in sync.

Eventually I plan to move this inside the application, as we can track everything
in the log file then